requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.2
scikit-learn==1.3.2
tenacity==8.2.3

# Machine Learning
xgboost==2.0.3
lightgbm==4.1.0
prophet==1.1.5

# API Client
requests==2.31.0
python-dateutil==2.8.2

# Apache Airflow
apache-airflow==2.8.0
apache-airflow-providers-http==4.7.0

# MLflow & Experiment Tracking
mlflow==2.9.2
dagshub==0.3.15

# DVC
dvc==3.36.1
dvc-s3==3.0.1

# FastAPI & Web Service
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Data Quality & Profiling
ydata-profiling==4.6.4
great-expectations==0.18.8

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Code Quality
black==23.12.1
pylint==3.0.3
flake8==7.0.0
isort==5.13.2

# Utilities
boto3==1.34.14
minio==7.2.3
matplotlib==3.8.2
seaborn==0.13.1
scipy==1.11.4

# Drift Detection
alibi-detect==0.11.4
evidently==0.4.11
